Add V6 and V7
In V6 ROC across the different datasets is explored. In V7 I also compare more indetails with table of truth where the diagnostic could be improved.

Finally I discover that I can actually model everything using binomial glm (!). I thin this is the way to go for comparign the different criteria and groupings.

One main THM is that, in Wards data,, 13.06 % of the subjects are either in the wrong group (ctl vs syn) or do not score "high enough" to the qeustionnaire.